Should be ready for Ubuntu 22.04
Works with Ubuntu 22.04 in preliminary test with 3 nodes. I got hung up on trying to make this work with Fedora CoreOS (FCOS) because I assumed that doing so would be better for running kubernetes (in particular k3s). As it turns out, running fwknop on FCOS requires firewalld, and that causes problems for both FCOS and potentially k3s. Meanwhile the k3s docs seem to indicate that k3s will work with Ubuntu 22.04.